+++ misc/metrics/examples/metrics/main.rs
@@ -51,11 +51,17 @@
 use env_logger::Env;
 use futures::executor::block_on;
 use futures::stream::StreamExt;
-use libp2p::core::Multiaddr;
-use libp2p::metrics::{Metrics, Recorder};
-use libp2p::swarm::{NetworkBehaviour, SwarmEvent};
-use libp2p::{identify, identity, ping, PeerId, Swarm};
+use libp2p_core::{identity, upgrade::Version, Multiaddr, PeerId, Transport};
+use libp2p_identify as identify;
+use libp2p_metrics::{Metrics, Recorder};
+use libp2p_noise as noise;
+use libp2p_ping as ping;
 use libp2p_swarm::keep_alive;
+use libp2p_swarm::NetworkBehaviour;
+use libp2p_swarm::Swarm;
+use libp2p_swarm::SwarmEvent;
+use libp2p_tcp as tcp;
+use libp2p_yamux as yamux;
 use log::info;
 use prometheus_client::registry::Registry;
 use std::error::Error;
@@ -69,10 +75,14 @@ fn main() -> Result<(), Box<dyn Error>> {
     let local_key = identity::Keypair::generate_ed25519();
     let local_peer_id = PeerId::from(local_key.public());
     let local_pub_key = local_key.public();
-    info!("Local peer id: {:?}", local_peer_id);
+    info!("Local peer id: {local_peer_id:?}");
 
     let mut swarm = Swarm::without_executor(
-        block_on(libp2p::development_transport(local_key))?,
+        tcp::async_io::Transport::default()
+            .upgrade(Version::V1)
+            .authenticate(noise::NoiseAuthenticated::xx(&local_key)?)
+            .multiplex(yamux::YamuxConfig::default())
+            .boxed(),
         Behaviour::new(local_pub_key),
         local_peer_id,
     );
@@ -115,6 +125,7 @@ fn main() -> Result<(), Box<dyn Error>> {
 /// For illustrative purposes, this includes the [`keep_alive::Behaviour`]) behaviour so the ping actually happen
 /// and can be observed via the metrics.
 #[derive(NetworkBehaviour)]
+#[behaviour(prelude = "libp2p_swarm::derive_prelude")]
 struct Behaviour {
     identify: identify::Behaviour,
     keep_alive: keep_alive::Behaviour,
@@ -122,7 +133,7 @@ struct Behaviour {
 }
 
 impl Behaviour {
-    fn new(local_pub_key: libp2p::identity::PublicKey) -> Self {
+    fn new(local_pub_key: identity::PublicKey) -> Self {
         Self {
             ping: ping::Behaviour::default(),
             identify: identify::Behaviour::new(identify::Config::new(