+++ protocols/gossipsub/tests/smoke.rs
@@ -29,16 +29,16 @@ use std::{
 };
 
 use futures::StreamExt;
-use libp2p::core::{
+use libp2p_core::{
     identity, multiaddr::Protocol, transport::MemoryTransport, upgrade, Multiaddr, Transport,
 };
-use libp2p::gossipsub::{
+use libp2p_gossipsub::{
     Gossipsub, GossipsubConfigBuilder, GossipsubEvent, IdentTopic as Topic, MessageAuthenticity,
     ValidationMode,
 };
-use libp2p::plaintext::PlainText2Config;
-use libp2p::swarm::{Swarm, SwarmEvent};
-use libp2p::yamux;
+use libp2p_plaintext::PlainText2Config;
+use libp2p_swarm::{Swarm, SwarmEvent};
+use libp2p_yamux as yamux;
 
 struct Graph {
     pub nodes: Vec<(Multiaddr, Swarm<Gossipsub>)>,
@@ -216,8 +216,7 @@ fn multi_hop_propagation() {
         });
         if !all_subscribed {
             return TestResult::error(format!(
-                "Timed out waiting for all nodes to subscribe but only have {:?}/{:?}.",
-                subscribed, num_nodes,
+                "Timed out waiting for all nodes to subscribe but only have {subscribed:?}/{num_nodes:?}.",
             ));
         }
 
@@ -246,8 +245,7 @@ fn multi_hop_propagation() {
         });
         if !all_received {
             return TestResult::error(format!(
-                "Timed out waiting for all nodes to receive the msg but only have {:?}/{:?}.",
-                received_msgs, num_nodes,
+                "Timed out waiting for all nodes to receive the msg but only have {received_msgs:?}/{num_nodes:?}.",
             ));
         }