+++ protocols/identify/src/lib.rs
@@ -46,22 +46,22 @@ pub use self::protocol::{Info, UpgradeError, PROTOCOL_NAME, PUSH_PROTOCOL_NAME};
 
 #[deprecated(
     since = "0.40.0",
-    note = "Use re-exports that omit `Identify` prefix, i.e. `libp2p::identify::Config`"
+    note = "Use re-exports that omit `Identify` prefix, i.e. `libp2p_identify::Config`"
 )]
 pub type IdentifyConfig = Config;
 
 #[deprecated(
     since = "0.40.0",
-    note = "Use re-exports that omit `Identify` prefix, i.e. `libp2p::identify::Event`"
+    note = "Use re-exports that omit `Identify` prefix, i.e. `libp2p_identify::Event`"
 )]
 pub type IdentifyEvent = Event;
 
-#[deprecated(since = "0.40.0", note = "Use libp2p::identify::Behaviour instead.")]
+#[deprecated(since = "0.40.0", note = "Use libp2p_identify::Behaviour instead.")]
 pub type Identify = Behaviour;
 
 #[deprecated(
     since = "0.40.0",
-    note = "Use re-exports that omit `Identify` prefix, i.e. `libp2p::identify::Info`"
+    note = "Use re-exports that omit `Identify` prefix, i.e. `libp2p_identify::Info`"
 )]
 pub type IdentifyInfo = Info;